mv process listing to ResourceArea + use RA.getNames in processListItem
leaving the process listing in Process would result in the following:
- Process imports RA to get the name of the resource IProcess,
- RA/Module imports Process/Module as it has done before.
-> circular dependency.

Within this commit RA.getNames also immediately gets used
ProcessListItem, so ca0dc77 gets "undone".
