+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Process/Module.ts
@@ -16,7 +16,5 @@ export var register = (angular) => {
         .directive("adhWorkflowSwitch", [
             "adhConfig", "adhHttp", "adhPermissions", "$q", "$translate", "$window", AdhProcess.workflowSwitchDirective])
         .directive("adhProcessView", ["adhTopLevelState", "adhProcess", "$compile", AdhProcess.processViewDirective])
-        .directive("adhProcessListItem", ["adhConfig", "adhHttp", "adhProcess", AdhProcess.listItemDirective])
-        .directive("adhProcessListing", ["adhConfig", "adhHttp", AdhProcess.listingDirective])
         .directive("adhCurrentProcessTitle", ["adhTopLevelState", "adhHttp", AdhProcess.currentProcessTitleDirective]);
 };
