+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Process/Process.ts
@@ -6,23 +6,11 @@ import * as AdhPermissions from "../Permissions/Permissions";
 import * as AdhUtil from "../Util/Util";
 import * as AdhTopLevelState from "../TopLevelState/TopLevelState";
 
-import * as SIDescription from "../../Resources_/adhocracy_core/sheets/description/IDescription";
-import * as SIImageReference from "../../Resources_/adhocracy_core/sheets/image/IImageReference";
-import * as SILocationReference from "../../Resources_/adhocracy_core/sheets/geo/ILocationReference";
-import * as SIName from "../../Resources_/adhocracy_core/sheets/name/IName";
 import * as SITitle from "../../Resources_/adhocracy_core/sheets/title/ITitle";
 import * as SIWorkflow from "../../Resources_/adhocracy_core/sheets/workflow/IWorkflowAssignment";
-import RIProcess from "../../Resources_/adhocracy_core/resources/process/IProcess";
 
 var pkgLocation = "/Process";
 
-
-// mirrors adhocracy_core.sheets.workflow.StateData
-export interface IStateData {
-    name : string;
-    description : string;
-    start_date : string;
-}
 
 export interface IProcessProperties {
     hasCreatorParticipate? : boolean;
@@ -38,19 +26,6 @@ export interface IProcessProperties {
     proposalSheet?;
     proposalVersionClass;
 }
-
-export var getStateData = (sheet : SIWorkflow.Sheet, name : string) : IStateData => {
-    for (var i = 0; i < sheet.state_data.length; i++) {
-        if (sheet.state_data[i].name === name) {
-            return sheet.state_data[i];
-        }
-    }
-    return {
-        name: null,
-        description: null,
-        start_date: null
-    };
-};
 
 
 export class Provider implements angular.IServiceProvider {
@@ -149,55 +124,6 @@ export var processViewDirective = (
                     element.html(template);
                     $compile(element.contents())(scope);
                 }
-            });
-        }
-    };
-};
-
-export var listItemDirective = (
-    adhConfig : AdhConfig.IService,
-    adhHttp : AdhHttp.Service,
-    adhProcess : Service
-) => {
-    return {
-        restrict: "E",
-        templateUrl: adhConfig.pkg_path + pkgLocation + "/ListItem.html",
-        scope: {
-            path: "@"
-        },
-        link: (scope) => {
-            adhHttp.get(scope.path).then((process) => {
-                if (process.data[SIImageReference.nick] && process.data[SIImageReference.nick].picture) {
-                    scope.picture = process.data[SIImageReference.nick].picture;
-                }
-                scope.title = process.data[SITitle.nick].title;
-                scope.processName = adhProcess.getName(process.content_type);
-                if (process.data[SILocationReference.nick] && process.data[SILocationReference.nick].location) {
-                    adhHttp.get(process.data[SILocationReference.nick].location).then((loc) => {
-                        scope.locationText = loc.data[SITitle.nick].title;
-                    });
-                }
-                var workflow = process.data[SIWorkflow.nick];
-                scope.participationStartDate = getStateData(workflow, "participate").start_date;
-                scope.participationEndDate = getStateData(workflow, "evaluate").start_date;
-                scope.shortDesc = process.data[SIDescription.nick].short_description;
-            });
-        }
-    };
-};
-
-export var listingDirective = (adhConfig : AdhConfig.IService, adhHttp : AdhHttp.Service) => {
-    return {
-        restrict: "E",
-        scope: {},
-        templateUrl: adhConfig.pkg_path + pkgLocation + "/Listing.html",
-        link: (scope) => {
-            scope.contentType = RIProcess.content_type;
-            scope.params = {
-                depth: "all"
-            };
-            adhHttp.get("/").then((root) => {
-                scope.overviewDesc = root.data[SIDescription.nick].description;
             });
         }
     };
